Initial stm32 common i2c based on stm32f4x
I2C customized routines copied from STM32Cube HAL; Mynewt HAL only uses
custom routines now
